+++ store/iavl/tree.go
@@ -33,6 +33,7 @@ type (
 		GetImmutable(version int64) (*iavl.ImmutableTree, error)
 		SetInitialVersion(version uint64)
 		Iterator(start, end []byte, ascending bool) (types.Iterator, error)
+		LoadVersionForOverwriting(targetVersion int64) (int64, error)
 	}
 
 	// immutableTree is a simple wrapper around a reference to an iavl.ImmutableTree
@@ -91,14 +92,6 @@ func (it *immutableTree) GetImmutable(version int64) (*iavl.ImmutableTree, error
 	return it.ImmutableTree, nil
 }
 
-func (it *immutableTree) AvailableVersions() []int {
-	return []int{}
-}
-
-func (it *immutableTree) LoadVersionForOverwriting(targetVersion int64) error {
+func (it *immutableTree) LoadVersionForOverwriting(targetVersion int64) (int64, error) {
 	panic("cannot call 'LoadVersionForOverwriting' on an immutable IAVL tree")
-}
-
-func (it *immutableTree) WorkingHash() []byte {
-	panic("cannot call 'WorkingHash' on an immutable IAVL tree")
 }